requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
jinja2>=3.1.0

# NLP / ML
spacy>=3.7.0
sentence-transformers>=3.0.0
transformers>=4.41.0
torch>=2.3.0

# Data
scikit-learn>=1.5.0
networkx>=3.3
numpy>=1.26.0
pandas>=2.2.0

# Utilities
feedparser>=6.0.0
rich>=13.7.0
